Types for buildVariables function
